requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1.0
psycopg[binary]>=3.1
python-dotenv>=1.0
cryptography>=42.0
gunicorn>=21.2
email-validator>=2.1.0